refactor: move WeatherList model to improve project structure
- Moved WeatherList model from core/network/response/ to models/ for improved code organization.
- Updated all relevant imports across services, view models, and tests to reference the new path.
- Ensures better separation of concerns by keeping models centralized under the models/ directory.

This refactor enhances code maintainability and aligns with project structure best practices.
