+++ test/repositories/weather_repository_impl_test.dart
@@ -5,10 +5,10 @@ import 'package:weather_app/core/logger/logger_provider.dart';
 import 'package:weather_app/core/network/api_error.dart';
 import 'package:weather_app/core/network/dio_error_handler.dart';
 import 'package:weather_app/core/network/response/result.dart';
-import 'package:weather_app/core/network/response/weather_list.dart';
 import 'package:weather_app/core/network/response/weather_response.dart';
 import 'package:weather_app/models/city_name.dart';
 import 'package:weather_app/models/weather_description.dart';
+import 'package:weather_app/models/weather_list.dart';
 import 'package:weather_app/models/weather_main.dart';
 import 'package:weather_app/models/weather_wind.dart';
 import 'package:weather_app/repositories/weather_repository_impl.dart';
