+++ test/view_model/city_search_view_model_test.dart
@@ -3,10 +3,10 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:weather_app/core/network/api_error.dart';
 import 'package:weather_app/core/network/response/result.dart';
-import 'package:weather_app/core/network/response/weather_list.dart';
 import 'package:weather_app/core/network/response/weather_response.dart';
 import 'package:weather_app/models/city_name.dart';
 import 'package:weather_app/models/weather_description.dart';
+import 'package:weather_app/models/weather_list.dart';
 import 'package:weather_app/models/weather_main.dart';
 import 'package:weather_app/models/weather_wind.dart';
 import 'package:weather_app/repositories/weather_repository_provider.dart';
